pyproject.toml
[tool.poetry]
name = "dbb-software-task"
version = "0.1.0"
description = ""
authors = ["Alexandr Zhydyk <zhydykalex@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
fastapi-pagination = "^0.12.31"
fastapi-filter = {extras = ["sqlalchemy"], version = "^2.0.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.6.9"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "**/migrations/*.py",
    ".git",
    "__pycache__/",
    ".idea/",
    ".log",
]


# Same as Black.
line-length = 120
indent-width = 4

target-version = "py312"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.params.Depends",
    "fastapi.Query",
    "fastapi.params.Query",
    "fastapi.Security",
    "fastapi_filter.FilterDepends"
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
select = ["A", "B", "E", "F", "UP", "I001"]
#ignore = ["A003", "E402"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = [
    "E", "F",
    "F401", # Remove unused imports.
    "I001" # Import block is un-sorted or un-formatted
]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
